+++ core/lib/sn/Areas.d.ts
@@ -1,8 +1,9 @@
 import { IAreas } from './CmnInterface';
 export declare class Areas implements IAreas {
     hAreas: {
-        [name: string]: number;
+        [name: number]: number;
     };
+    clear(): void;
     search(idx: number): boolean;
     record(idx: number): void;
     erase(idx: number): void;